Run vertex-wise linear mixed model using lme4::lmer()

This is is the main function in v0 of verywise. It runs the single_lmm function across all vertices.

Usage

run_vw_lmm(
  formula,
  subj_dir,
  outp_dir = NULL,
  pheno = NULL,
  hemi = c("both", "lh", "rh"),
  seed = 3108,
  ...
)

Arguments

formula

: model formula object (this should specify a LME model)

subj_dir

: path to the FreeSurfer data, this expects a verywise structure.

outp_dir

: output path, where do you want results to be stored. If none is provided by the user, a "results" sub-directory will created inside subj_dir.

pheno

: the phenotype data or a path to the data file.

hemi

: (default = "both") which hemispheres to run.

seed

: (default = 3108) random seed.

...

Arguments passed on to hemi_vw_lmm

measure

: (default = gsub("vw_", "", all.vars(formula)[1])) vertex-wise measure this should be the same as specified in formula

fwhm

: (default = 10) full-width half maximum value

target

: (default = "fsaverage") template on which to register vertex-wise data.

model

: (default = "lme4::lmer") # "stats::lm"

Value

A list of file-backed matrices containing pooled coefficients, SEs, t- and p- values and residuals.

Author

Serena Defina, 2024.