Build "supersubject" by stacking all vertex data in one large file-backed matrix with dimensions n_subjects x n_vertices.

Build "supersubject" by stacking all vertex data in one large file-backed matrix with dimensions n_subjects x n_vertices.

Usage

build_supersubject(
  subj_dir,
  folder_id,
  files_list = list.dirs.till(subj_dir, n = 2),
  measure = "thickness",
  hemi = "lh",
  fwhmc = "fwhm10",
  target = "fsaverage",
  backing = file.path(subj_dir, paste0(hemi, ".", measure, ".supersubject.bk")),
  mask = TRUE,
  save_rds = FALSE,
  cluster,
  verbose = TRUE
)

Arguments

subj_dir

: path to the FreeSurfer data, this expects a verywise structure.

folder_id

: column in phenotype file that holds the IDs of the observations to include. This also expects a verywise format (e.g. "site1/sub-1_ses-01")

files_list

: (default = list.dirs.till(subj_dir, n = 2)) allows to specify manually which folders or files to include. You would typically not use this parameter, but rather use folder_id to specify which observations to include.

measure

: (default = "thickness"), vertex-wise measure, used to identify files.

hemi

: (default = "lh") hemisphere, used to identify files.

fwhmc

: (default = "fwhm10") full-width half maximum value, used to identify files.

target

: (default = "fsaverage"), used to identify files.

backing

: (default = subj_dir) location to save the matrix backingfile.

mask

: (dafault = TRUE) only keep cortical vertices, according to FreeSurfer cortical map.

save_rds

: (default = FALSE) save the supersubject file metadata for re-use in other sessions.

cluster

: the parallel cluster

verbose

: (default = TRUE)

Value

A Filebacked Big Matrix with vertex data for all subjects (dimensions: n_subjects x n_vertices)

Author

Serena Defina, 2024.