Simulate FreeSurfer data

Simulating FreeSurfer data for multiple cohorts/sites and multiple timepoints/sessions. This function emulates the folder structure output obtained by calling FreeSurfer recon_all command. This follows the folder names "sub-", participant ID, "_ses-", session ID.

Usage

simulate_freesurfer_data(
  path,
  data_structure = list(cohort1 = list(sessions = c("01", "02"), n_subjects = 100),
    cohort2 = list(sessions = c("01", "02"), n_subjects = 150)),
  vw_resolution = 163842,
  measure = "thickness",
  hemi = "lh",
  fwhmc = "fwhm10",
  target = "fsaverage",
  vw_mean = 6.5,
  vw_sd = 0.5,
  simulate_association = NULL,
  seed = 31081996
)

Arguments

path

Where should the data be created.

data_structure

A nested list, with top level determining the cohort/ dataset/site. Each site is itself a list with two items: "sessions": a vector of session names/numbers; and "n_subjects": an integer indicating the number of subjects.

vw_resolution

(default = 163842) data dimension (how many vertices)

measure

(default = "thickness"), measure, used in file names.

hemi

(default = "lh") hemisphere, used in file names.

fwhmc

(default = "fwhm10") full-width half maximum value, used in file names.

target

(default = "fsaverage"), used in file names.

vw_mean

(default = 6.5) mean of the simulated vertex-wise data.

vw_sd

(default = 0.5) standard deviation of the simulated vertex-wise data.

simulate_association

(default = NULL) simulate an association in the format beta * variable. This is by default isolated to three regions: the superior temporal gyrus, precentral gyrus and middle temporal gyrus.

seed

(default = 31081996) seed used for randomization.

Author

Serena Defina, 2024.